6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ long int:c;
 long long:d;
 float:f;
 
-printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("size of a int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("size of a long long: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("size of a char: %d byte(s)\n", sizeof(a));
+printf("size of a int: %d byte(s)\n", sizeof(b));
+printf("size of a long int: %d byte(s)\n", sizeof(c));
+printf("size of a long long: %d byte(s)\n", (sizeof(d));
+printf("size of a float: %d byte(s)\n", sizeof(f));
 return (0);
 }
